alex/store-blob-storage: renames
